Se termina el funcionamiento del sistema
