+++ resources/js/errores.js
@@ -23,3 +23,14 @@ export const notificarErroresTelefono = ( errores ) => {
         document.querySelector(elementoError).innerHTML =errores[clave][0];
     }
 }
+
+
+// metodo para notificar errores en el formulario al editar usuario
+export const notificarErroresEditarUsuario = ( errores ) => {
+    for (let clave in errores){
+        const elemento = `#${clave}-edit`; 
+        const elementoError = `#${clave}-validate-edit`;
+        document.querySelector(elemento).classList.add('is-invalid');
+        document.querySelector(elementoError).innerHTML =errores[clave][0];
+    }
+}
