+++ resources/views/agenda.blade.php
@@ -20,6 +20,7 @@
         @include('agregar')
         @include('modalVerTelefonos')
         @include('modalAgregarTelefono')
+        @include('modalEditarUsuario')
     </body>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.1/dist/js/bootstrap.bundle.min.js" integrity="sha384-/bQdsTh/da6pkI1MST/rWKFNjaCP5gBSY4sEBT38Q/9RBh9AH40zEOg7Hlq2THRZ" crossorigin="anonymous"></script>
     <script src="{{ asset('js/app.js') }}" defer></script>
